# api/openapi.yaml
openapi: 3.0.3
info: {title: Netlab API, version: 1.0.0}
servers: [{url: /api/v1}]
security: [{bearer: []}]
paths:
  /templates/{id}/cache:
    parameters: [{$ref: '#/components/parameters/Id'}]
    get:
      operationId: templateCache
      responses:
        '200': {description: Local base disk cache, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/TemplateCache'}}}}}
    delete:
      operationId: trimTemplateCache
      responses: {'202': {$ref: '#/components/responses/Accepted'}}
  /identity/password:
    put:
      operationId: changePassword
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ChangePassword'}}}}
      responses: {'204': {description: Password changed; sign in again}}
  /principals/{id}/grants:
    parameters: [{$ref: '#/components/parameters/Id'}]
    get:
      operationId: principalGrants
      responses:
        '200': {description: Account grants, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/ScopeGrant'}}}}}
  /nodes/{id}/storage-device:
    parameters: [{$ref: '#/components/parameters/Id'}]
    get:
      operationId: nodeStorageDevices
      responses:
        '200': {description: Physical disks, selected Ceph disk and preparation task, content: {application/json: {schema: {$ref: '#/components/schemas/NodeStorageDevices'}}}}
    put:
      operationId: configureNodeStorage
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ConfigureNodeStorage'}}}}
      responses:
        '202': {$ref: '#/components/responses/Accepted'}
  /storage-pools/{id}/ceph:
    parameters: [{$ref: '#/components/parameters/Id'}]
    get:
      operationId: cephStatus
      responses:
        '200': {description: Live managed Ceph status, content: {application/json: {schema: {$ref: '#/components/schemas/CephStatus'}}}}
    put:
      operationId: configureCephPool
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ConfigureCephPool'}}}}
      responses:
        '202': {$ref: '#/components/responses/Accepted'}
  /storage-pools/{id}/assets:
    parameters: [{$ref: '#/components/parameters/Id'}]
    get:
      operationId: storagePoolAssets
      responses:
        '200': {description: Assets using this pool, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/StoragePoolAsset'}}}}}
  /environments/{id}/assets/{assetId}/migrations:
    parameters:
      - {$ref: '#/components/parameters/Id'}
      - {name: assetId, in: path, required: true, schema: {type: string}}
    get:
      operationId: migrationDestinations
      responses:
        '200': {description: Compatible migration destinations, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/MigrationDestination'}}}}}
    post:
      operationId: migrateAsset
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/MigrationRequest'}}}}
      responses:
        '202': {$ref: '#/components/responses/Accepted'}
        '409': {description: Revision changed or an environment operation is unfinished}
  /volumes:
    get:
      operationId: listVolumes
      responses:
        '200': {description: Persistent volumes, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/PersistentVolume'}}}}}
    post:
      operationId: createVolume
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CreateVolume'}}}}
      responses:
        '202': {description: Volume creation accepted, content: {application/json: {schema: {$ref: '#/components/schemas/Operation'}}}}
  /volumes/{id}:
    parameters: [{$ref: '#/components/parameters/Id'}]
    get:
      operationId: getVolume
      responses:
        '200': {description: Persistent volume, content: {application/json: {schema: {$ref: '#/components/schemas/PersistentVolume'}}}}
        '404': {description: Volume not found}
    put:
      operationId: resizeVolume
      requestBody: {required: true, content: {application/json: {schema: {type: object, required: [sizeGiB], properties: {sizeGiB: {type: integer, format: int64, minimum: 1}}}}}}
      responses:
        '202': {description: Detached volume expansion accepted, content: {application/json: {schema: {$ref: '#/components/schemas/Operation'}}}}
        '409': {description: Volume is attached or an operation is pending}
    delete:
      operationId: deleteVolume
      responses:
        '202': {description: Unreferenced volume deletion accepted, content: {application/json: {schema: {$ref: '#/components/schemas/Operation'}}}}
        '409': {description: Volume is referenced}
  /environments/{id}/traffic:
    parameters: [{$ref: '#/components/parameters/Id'}]
    get:
      operationId: observeTraffic
      responses:
        '200': {description: Rolling sampled conversations; counters are statistical estimates, content: {application/json: {schema: {$ref: '#/components/schemas/TrafficObservation'}}}}
  /environments/{id}/captures:
    parameters: [{$ref: '#/components/parameters/Id'}]
    get:
      operationId: listCaptures
      responses:
        '200': {description: Node capture segments and unavailable nodes, content: {application/json: {schema: {$ref: '#/components/schemas/CaptureList'}}}}
    post:
      operationId: startCapture
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CreateCapture'}}}}
      responses:
        '201': {description: Actual per-node start results, content: {application/json: {schema: {$ref: '#/components/schemas/CaptureList'}}}}
  /environments/{id}/captures/{nodeId}/{captureId}:
    parameters:
      - {$ref: '#/components/parameters/Id'}
      - {name: nodeId, in: path, required: true, schema: {type: string}}
      - {name: captureId, in: path, required: true, schema: {type: string}}
    get:
      operationId: captureFlows
      responses:
        '200': {description: Observed conversations and counters, content: {application/json: {schema: {$ref: '#/components/schemas/CaptureDetail'}}}}
    post:
      operationId: stopCapture
      responses:
        '200': {description: Stopped segment, content: {application/json: {schema: {$ref: '#/components/schemas/CaptureSegment'}}}}
    delete:
      operationId: deleteCapture
      responses:
        '204': {description: Capture and its file removed}
  /environments/{id}/captures/{nodeId}/{captureId}/file:
    parameters:
      - {$ref: '#/components/parameters/Id'}
      - {name: nodeId, in: path, required: true, schema: {type: string}}
      - {name: captureId, in: path, required: true, schema: {type: string}}
    get:
      operationId: downloadCapture
      responses:
        '200': {description: Completed PCAPNG stream, content: {application/octet-stream: {schema: {type: string, format: binary}}}}
  /system/update:
    get:
      operationId: systemUpdate
      responses:
        '200': {description: Installed version and update progress; administrator only, content: {application/json: {schema: {$ref: '#/components/schemas/SystemUpdate'}}}}
    post:
      operationId: applySystemUpdate
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ApplySystemUpdate'}}}}
      responses:
        '202': {description: Update accepted, content: {application/json: {schema: {$ref: '#/components/schemas/SystemUpdate'}}}}
        '409': {description: Update already running or version changed}
...
  /backup-repositories:
    get:
      operationId: listBackupRepositories
      responses:
        '200': {description: Backup repositories, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/BackupRepository'}}}}}
    post:
      operationId: createBackupRepository
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CreateBackupRepository'}}}}
      responses:
        '201': {description: Repository connection queued, content: {application/json: {schema: {$ref: '#/components/schemas/BackupRepository'}}}}
  /backup-repositories/{id}:
    parameters: [{$ref: '#/components/parameters/Id'}]
    delete:
      operationId: deleteBackupRepository
      responses:
        '204': {description: Registration removed; repository data retained}
        '409': {description: Repository contains backups or an active task}
  /backup-repositories/{id}/credentials:
    parameters: [{$ref: '#/components/parameters/Id'}]
    get:
      operationId: backupRepositoryCredentials
      responses:
        '200': {description: Repository credentials; administrator only, content: {application/json: {schema: {$ref: '#/components/schemas/BackupCredentials'}}}}
  /backup-repositories/{id}/refresh:
    parameters: [{$ref: '#/components/parameters/Id'}]
    post:
      operationId: refreshBackupRepository
      responses:
        '202': {$ref: '#/components/responses/Accepted'}
  /backup-repositories/{repositoryId}/backups:
    parameters: [{name: repositoryId, in: path, required: true, schema: {type: string}}]
    get:
      operationId: listRepositoryBackups
      parameters: [{$ref: '#/components/parameters/Cursor'}, {$ref: '#/components/parameters/Limit'}]
      responses:
        '200': {description: Repository catalog; administrator only, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/BackupSummary'}}}}}
  /backup-repositories/{repositoryId}/backups/{backupId}:
    parameters:
      - {name: repositoryId, in: path, required: true, schema: {type: string}}
      - {name: backupId, in: path, required: true, schema: {type: string}}
    delete:
      operationId: deleteRepositoryBackup
      responses:
        '202': {$ref: '#/components/responses/Accepted'}
  /environments/{id}/backups:
    parameters: [{$ref: '#/components/parameters/Id'}]
    get:
      operationId: listBackups
      parameters: [{$ref: '#/components/parameters/Cursor'}, {$ref: '#/components/parameters/Limit'}]
      responses:
        '200': {description: Environment backups, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/BackupSummary'}}}}}
    post:
      operationId: createBackup
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CreateBackup'}}}}
      responses:
        '201': {description: Backup queued, content: {application/json: {schema: {$ref: '#/components/schemas/BackupSummary'}}}}
  /environments/{id}/backups/{backupId}:
    parameters:
      - {$ref: '#/components/parameters/Id'}
      - {name: backupId, in: path, required: true, schema: {type: string}}
    delete:
      operationId: deleteBackup
      responses:
        '202': {$ref: '#/components/responses/Accepted'}
  /environments/{id}/backups/{backupId}/restore:
    parameters:
      - {$ref: '#/components/parameters/Id'}
      - {name: backupId, in: path, required: true, schema: {type: string}}
    post:
      operationId: restoreBackup
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/RestoreRecoveryPoint'}}}}
      responses:
        '202': {$ref: '#/components/responses/Accepted'}
  /environments/{id}/recovery-points:
    parameters: [{$ref: '#/components/parameters/Id'}]
    get:
      operationId: listRecoveryPoints
      parameters: [{$ref: '#/components/parameters/Cursor'}, {$ref: '#/components/parameters/Limit'}]
      responses:
        '200': {description: Environment recovery points, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/RecoveryPointSummary'}}}}}
    post:
      operationId: captureRecoveryPoint
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CaptureRecoveryPoint'}}}}
      responses:
        '201':
          description: Capture queued; normal shutdown and original power state restoration are part of the operation
          headers:
            Operation-Location: {schema: {type: string}}
          content: {application/json: {schema: {$ref: '#/components/schemas/RecoveryPointSummary'}}}
        '409': {description: Revision changed or another environment operation is active}
  /environments/{id}/recovery-points/{pointId}:
    parameters:
      - {$ref: '#/components/parameters/Id'}
      - {name: pointId, in: path, required: true, schema: {type: string}}
    delete:
      operationId: deleteRecoveryPoint
      responses:
        '202': {$ref: '#/components/responses/Accepted'}
  /environments/{id}/recovery-points/{pointId}/restore:
    parameters:
      - {$ref: '#/components/parameters/Id'}
      - {name: pointId, in: path, required: true, schema: {type: string}}
    post:
      operationId: restoreRecoveryPoint
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/RestoreRecoveryPoint'}}}}
      responses:
        '202': {$ref: '#/components/responses/Accepted'}
  /storage-pools:
    get:
      operationId: listStoragePools
      responses:
        '200': {description: Storage pools and current capacity, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/StoragePool'}}}}}
    post:
      operationId: createStoragePool
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CreateStoragePool'}}}}
      responses:
        '201': {description: Registered storage, content: {application/json: {schema: {$ref: '#/components/schemas/StoragePool'}}}}
  /storage-pools/{id}:
    parameters: [{$ref: '#/components/parameters/Id'}]
    delete:
      operationId: deleteStoragePool
      responses:
        '202': {$ref: '#/components/responses/Accepted'}
        '409': {description: Pool is referenced, content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}}
  /environments/{id}/assets/{assetId}/templates:
    parameters:
      - {$ref: '#/components/parameters/Id'}
      - {name: assetId, in: path, required: true, schema: {type: string}}
    post:
      operationId: captureAssetTemplate
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CaptureTemplate'}}}}
      responses:
        '201':
          description: Template capture queued; source VM remains stopped
          headers:
            Operation-Location: {schema: {type: string}, description: Operation resource}
          content: {application/json: {schema: {$ref: '#/components/schemas/Template'}}}
  /nodes/{id}/retirement:
    parameters: [{$ref: '#/components/parameters/Id'}]
    get:
      operationId: nodeRetirement
      responses:
        '200': {description: Node exit plan and progress, content: {application/json: {schema: {$ref: '#/components/schemas/NodeRetirement'}}}}
    post:
      operationId: retireNode
      responses:
        '202': {$ref: '#/components/responses/Accepted'}
        '409': {description: Node has dependencies to resolve}
    delete:
      operationId: resumeNodeScheduling
      responses:
        '204': {description: Scheduling resumed}
  /nodes/{id}/interfaces:
    parameters: [{$ref: '#/components/parameters/Id'}]
    get:
      operationId: listNodeInterfaces
      responses:
        '200': {description: Current host interfaces, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/ExternalInterface'}}}}}
  /principals:
    get:
      operationId: listPrincipals
      parameters: [{$ref: '#/components/parameters/Cursor'}, {$ref: '#/components/parameters/Limit'}, {$ref: '#/components/parameters/Search'}, {name: kind, in: query, schema: {type: string, enum: [user, token]}}]
      responses:
        '200': {description: Accounts, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/Principal'}}}}}
    post:
      operationId: createUser
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CreateUser'}}}}
      responses:
        '201': {description: Created, content: {application/json: {schema: {$ref: '#/components/schemas/Principal'}}}}
  /principals/{id}:
    parameters: [{$ref: '#/components/parameters/Id'}]
    put:
      operationId: updateUser
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/UpdateUser'}}}}
      responses: {'204': {description: Updated}}
  /service-tokens:
    post:
      operationId: createServiceToken
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CreateServiceToken'}}}}
      responses:
        '201': {description: Token returned once, content: {application/json: {schema: {$ref: '#/components/schemas/IssuedServiceToken'}}}}
  /service-tokens/{id}:
    parameters: [{$ref: '#/components/parameters/Id'}]
    delete:
      operationId: revokeServiceToken
      responses: {'204': {description: Revoked}}
...
  /environments/{id}:
    parameters: [{$ref: '#/components/parameters/Id'}]
    get:
      operationId: getEnvironment
      responses:
        '200': {description: Environment, content: {application/json: {schema: {$ref: '#/components/schemas/Environment'}}}}
...
components:
  parameters:
    Id: {name: id, in: path, required: true, schema: {type: string}}
...
  responses:
    Accepted:
      description: Accepted
      content: {application/json: {schema: {$ref: '#/components/schemas/Operation'}}}
  securitySchemes:
    bearer: {type: http, scheme: bearer}
  schemas:
...
    ScopeGrant:
      type: object
      required: [scopeKind, scopeId, permissions]
      properties:
        scopeKind: {type: string, enum: [project, environment, asset]}
        scopeId: {type: string}
        permissions: {type: array, items: {$ref: '#/components/schemas/Permission'}}
...
    ChangePassword:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword: {type: string, format: password}
        newPassword: {type: string, minLength: 8, maxLength: 72, format: password}
...
    TemplateCache:
      type: object
      required: [nodeId, nodeName, bytes, reclaimable, reason]
      properties:
        nodeId: {type: string}
        nodeName: {type: string}
        bytes: {type: integer, format: int64}
        reclaimable: {type: boolean}
        reason: {type: string}
...
    ConfigureNodeStorage:
      type: object
      required: [device]
      properties:
        device: {type: string, description: Physical empty disk selected for Ceph; empty clears an unused selection}
...
    NodeStorageDevices:
      type: object
      required: [devices, selected]
      properties:
        devices: {type: array, items: {$ref: '#/components/schemas/StorageDevice'}}
        selected: {type: string}
        operation: {$ref: '#/components/schemas/Operation'}
...